src/ColorCheckerGenerator/requirements.txt
numpy
opencv-python
Pillow
reportlab
colour-science
